refactor: unify prompt preview component
